v1

+++ src/main/java/es/http/service/dao/IUsuarioDAO.java
@@ -3,7 +3,7 @@ package es.http.service.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import es.http.service.dto.Usuario;
 
-public interface IUsuarioDAO extends JpaRepository<Usuario, Long> {
+public interface IUsuarioDAO extends JpaRepository<Usuario, Integer> {
 
 	Usuario findByUsername(String username);
 }
